Fix account configuration value saving, add ServiceAccountModel
Previously, the configuration values were not synced to the database
correctly.  This patch ensures that not only can existing values
be updated, but new values can be written; and it uses a per-type
conversion function to ensure that values are read back correctly.
It also ensures that setting the display name of the account is the
last thing done prior to sync(), avoiding overwrite due to the
magic "name" configuration value.

The unit test is updated to ensure that no regressions can occur.

Finally, a ServiceAccountModel type is added, to allow per-service
account information to be displayed in a list view.